split SchemaGenerator and SchemaSanitizer
